requirements.txt
asgiref==3.11.0
attrs==25.4.0
coverage==7.12.0
dataclasses==0.6
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==38.2.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
librt==0.6.2
mypy==1.19.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
ruff==0.14.6
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0
